Closes #50
