+++ build.gradle
@@ -1,0 +1,28 @@
+///////////////////////////////////////////////////////////////////////////////////////
+//
+// INSPIRE ETS-REPOSITORY
+//
+///////////////////////////////////////////////////////////////////////////////////////
+// @author jonherrmann
+
+group = 'eu.inspire.validation'
+description = 'Repository for Executable Test Suites of the INSPIRE validator'
+
+apply plugin: 'groovy'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    compile 'org.codehaus.groovy:groovy-all:2.4.12'
+    testCompile group: 'junit', name: 'junit', version: '4.11'
+}
+
+sourceSets {
+    test {
+        groovy {
+            srcDirs = ['test/groovy']
+        }
+    }
+}
